requirements.txt
instabot
pandas
Pillow
praw
print-schema